requirements.txt
pishockpy
mastodon.py
